moved all funtionality to LeagueSnitch class
